+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 Added
   * AZ-1038: Add parameters to Key Vault module
+  * AZ-1070: Add `vm_backup_policy_type` new parameter
 
 Breaking
-  * AZ-1038: Bump Key Vault module. Use `var.keyvault_custom_name` to keep your Key Vault when upgrading.
+  * AZ-1038: Bump Key Vault module. Use `var.keyvault_custom_name` to keep your Key Vault when upgrading
+
+Fixed
+  * AZ-1070: Fix VM backup policy configuration when `vm_backup_policy_frequency` is set to `Weekly`
 
 # v7.2.2 - 2023-04-07